requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
distro==1.9.0
flatbuffers==24.3.25
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.42.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numba==0.59.1
numpy==1.26.4
onnxruntime==1.18.0
openai==1.33.0
opencv-python-headless==4.10.0.82
packaging==24.1
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.27.1
pydantic==2.7.4
pydantic_core==2.18.4
PyMatting==1.1.12
python-dotenv==1.0.1
referencing==0.35.1
rembg==2.0.57
requests==2.32.3
rpds-py==0.18.1
scikit-image==0.23.2
scipy==1.13.1
sniffio==1.3.1
sympy==1.12.1
tifffile==2024.5.22
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.1
uuid==1.30
yarl==1.9.4
